% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd_parser.R
\name{isd_parse}
\alias{isd_parse}
\title{Parse NOAA ISD/ISH data files}
\usage{
isd_parse(path, parallel = FALSE, cores = getOption("cl.cores", 2))
}
\arguments{
\item{path}{(character) file path. required}

\item{parallel}{(logical). do processing in parallel. Default: \code{FALSE}}

\item{cores}{(integer) number of cores to use: Default: 2. We look in
your option "cl.cores", but use default value if not found.}
}
\value{
A tibble (data.frame)
}
\description{
Parse NOAA ISD/ISH data files
}
\examples{
\dontrun{
path <- system.file('extdata/725300-94846-2014.gz', package = "isdparser")

(res <- isd_parse(path))

# in parallel
(out <- isd_parse(path, parallel = TRUE))

identical(res, out)
}
}
\references{
ftp://ftp.ncdc.noaa.gov/pub/data/noaa
}

