% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd_parse_csv.R
\name{isd_parse_csv}
\alias{isd_parse_csv}
\title{Parse NOAA ISD/ISH csv data files}
\usage{
isd_parse_csv(path)
}
\arguments{
\item{path}{(character) file path. required}
}
\value{
A tibble (data.frame)
}
\description{
Parse NOAA ISD/ISH csv data files
}
\details{
Note that the `rem` (remarks) and `eqd` columns are
not parsed, just as with [isd_parse()].
}
\section{Column information}{


- USAF MASTER and NCEI WBAN station identifiers are combined into an 11
character code with the column `station`
- Date and Time have been combined to the column `date`
- Call letter is synonymous with `call_sign` column
- WIND-OBSERVATION is abbreviated as column `wnd`
- SKY-CONDITION-OBSERVATION is abbreviated as column `cig`
- VISIBILITY-OBSERVATION is abbreviated as column `vis`
- AIR-TEMPERATURE-OBSERVATION air temperature is abbreviated as the column
header `tmp`
- AIR-TEMPERATURE-OBSERVATION dew point is abbreviated as the column
`dew`
- AIR-PRESSURE-OBSERVATION sea level pressure is abbreviated as the column
`slp`
}

\examples{
path <- system.file('extdata/00702699999.csv', package = "isdparser")
(res <- isd_parse_csv(path))
}
\references{
https://www.ncei.noaa.gov/data/global-hourly/access/
https://www.ncei.noaa.gov/data/global-hourly/doc/CSV_HELP.pdf
https://www.ncei.noaa.gov/data/global-hourly/doc/isd-format-document.pdf
}
